Implemented CURD For State Entity

+++ src/main/java/com/example/service/StateFetchService.java
@@ -3,6 +3,6 @@ package com.example.service;
 import com.example.entity.State;
 
 public interface StateFetchService {
-	State fetchById(int id);
+	State fetchById(Long id);
 
 }
