+++ src/main/java/com/example/service/StateUpdateService.java
@@ -3,6 +3,6 @@ package com.example.service;
 import com.example.entity.State;
 
 public interface StateUpdateService {
-	State updateData(State state , int id);
+	State updateData(State state , Long id);
 
 }
